test: split backtesting and feature caching suites
